Put speed and velocities fully into the movement class and changed velocities to ints. Made all game actors have reference to a movement object

+++ src/include/player.h
@@ -8,7 +8,7 @@
 class Player : public GameActor {
     public:
         Player(Controller* c);
-        Player(double x, double y, double w, double h, double vx, double vy,
+        Player(double x, double y, double w, double h, int vx, int vy,
                double speed, Controller* c, Weapon* wp);
         virtual void Update(double timeStep);
         virtual void UpdatePosition(double timeStep);
